Prompt to ignore unexecutable scripts
Keeps track of an ignoreableHookFiles set which is added to whenever the
ExecuteHookFiles function attempts to execute a script which does not have
the executable bit set. Users can type "y" to set the executable bit which
runs a chmod u+x on the script file or "n" to leave the permissions as-is
and add the hook file to the ignoreableHookFiles set. That hook file will be
silently ignored when responding to any subsequent events which might
otherwise have caused it to execute. The ignoring happens for the duration
of the session and is reset once Vim is closed and re-opened.