Crazy times with :ListVimHooks
:ListVimHooks opens another buffer that has it's own syntax coloring and
set of controls, emulated with silly while loop that traps user input by
continuousy calling getchar(). Pressing "x" on one of the VimHook lines will
toggle it's enabled/disabled state, which involves renaming the file and
updating all the internal data structures.

Scripts that don't have the execute bit set cause some confusing behavior.
In particular, if it's not disabled explicitly by adding ".disabled" to
the end of the script name, it'll still show up as "enabled" in the
:ListVimHooks output. Saying "n" at the prompt that asks if you want to set
the executable bit doesn't actually disable it, it just adds it to this
other internal dictionary of ignored VimHooks.