ListVimHooks: o, <cr> open in prev window
In fact I co-opted a bunch of code from NERDTree to implement this. Both "o"
and "<CR>" do the same thing -- try to open the relevant VimHook script in
the previous window. If the window contains a modified buffer or just can't
be overtaken because it's a special window (e.g., quickfix) then it will
open in a vertical split.

Still not sure what will happen in the exception cases.