Things seem to work with the new logic.
s:findHookFiles finds all the hook files and adds them to
s:patternBasedVimHooks AS WELL AS s:allVimHooks, which is just a master List
of all VimHooks. When s:executeHookFiles is called for the first time from
some file, the function will initialize the value of
s:vimHooksByFilename[filename] to a list of all VimHooks whose pattern
matches the filename. It THEN iterates through all VimHooks in that list to
find any which match the given event and exeute them.

:ListVimHooks still depends on the s:patternBasedVimHooks data structure.
May have to change this logic so we don't have duplicate data structures
hanging around.