Remove usage of patternBasedVimHooks
Instead just maintaining a List of all VimHooks. When we call
executeVimHooks() from a new file we will create a new sorted list of
VimHooks whose pattern matches that file's name and add it to the
vimHooksByFilename dictionary. Then we go through this list and fire of any
hooks which match the associated event. We only have to create this sorted
list once for each file. After that, it is just left in memory and can be
referenced in O(1) time.

Updated :ListVimHooks to use allVimHooks instead of patternBasedVimHooks.