Print warning when hooks return bad exit code
Uses the v:shell_error variable which should match the value of $? in your
shell. Error messages are concatenated together so they can all be printed
at once without being erased when the screen is redrawn.

Should perhaps expose a flag or something to disable these warning messages
if the user wants to ignore them.