Keep hook listing if user does not delete
When user types "d" on a line in the VimHook listing window, the plugin will
execute the shell command "rm -i HOOK_FILE". Afterwards, it now checks if
the file still exists (user may have entered "yes" or "no"). If it does not
exist, the corresponding line is removed from the listing window content.