Support debounce.wait and async options
Format of the debounce.wait option:

    vimhook.debounce.wait: 5

With this option set, the hook is wrapped in the bin/debounce script. The
full description of what bin/debounce does is contained in an earlier
commit.

Format of the async option (remmeber when just the option key is specified
the value is implicitly "true"):

    vimhook.async

With this option set, the hook is run with "&" tacked on the end, meaning
that in most shell environments this will fork the process. I'm not sure if
stdout and stderr can be successfully caught in this case. Will have to
test.