Allow vimhook.bufferoutput.feedkeys option
Explanation by example. If you have the following lines in your hook file

    vimhook.bufferoutput
    vimhook.bufferoutput.feedkeys = 100Gzt

then VimHooks will dump the stdout of the hook into a scratch buffer just
the same as before. After that's done it will type in whatever keys you
provided with the "feedkeys" option in Normal mode in that scratch buffer.
In this case it will go to line 100 and then put that line at the top of the
window.

The most common use case I have so far is just running "G" in the output
buffer to put the cursor at the bottom. This comes up when I'm writing unit
tests and want to see the latest tests at the bottom. Obviously you can get
as elaborate as you want with the Normal commands.