CreateSelfHook command
Proof of concept for creating a hook which is the file itself you're
editing. Provides an easier way to iterate using the Vim-as-REPL idea
without going and creating a whole new hook script and saving it to disk.
Because the current file does NOT follow the VimHook naming convention
however, it will not get automatically picked up when you start Vim.
CreateSelfHook will thus be the only way to add it to the list of current
hooks. Another way to think about this is that CreateSelfHook is a way of
manually adding the current file to the index of VimHooks, whether it
follows the VimHook file naming convention or not.